+++ src/lib/youtube-transcript-types.ts
@@ -1,0 +1,13 @@
+export interface TranscriptResponse {
+  text: string;
+  duration: number;
+  offset: number;
+  lang?: string;
+}
+
+export interface YoutubeTranscriptResponse {
+  videoId: string;
+  title?: string;
+  transcript: TranscriptResponse[];
+  error?: string;
+}
